requirements.txt
# CrewAI Idle Game Development Requirements

# Core CrewAI and AI dependencies
crewai>=0.28.0
crewai-tools>=0.1.6
langchain-google-genai>=1.0.1
google-generativeai>=0.3.2

# Environment and configuration
python-dotenv>=1.0.0

# GUI and game development
tkinter-tooltip>=2.0.0
Pillow>=10.0.0
pygame>=2.5.0

# Data handling and serialization
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Logging and utilities
colorlog>=6.7.0
rich>=13.0.0

# Testing framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Build and packaging
pyinstaller>=5.13.0
setuptools>=68.0.0
wheel>=0.41.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: Advanced features
numpy>=1.24.0
matplotlib>=3.7.0
requests>=2.31.0
